# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Your API",
    "description": "A simple example ASP.NET Core Web API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Article/type/{libelle}": {
      "get": {
        "tags": [
          "Article"
        ],
        "parameters": [
          {
            "name": "libelle",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArticleDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Fournisseur/{id}": {
      "get": {
        "tags": [
          "Fournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FournisseurDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Fournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FournisseurDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FournisseurDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FournisseurDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Fournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/Fournisseur/search": {
      "get": {
        "tags": [
          "Fournisseur"
        ],
        "parameters": [
          {
            "name": "ice",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviseLibelle",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/RefType/{id}": {
      "get": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RefType"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/RefTypeRetour": {
      "get": {
        "tags": [
          "RefTypeRetour"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeRetourDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeRetourDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RefTypeRetourDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RefTypeRetour"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RefTypeRetour/{id}": {
      "get": {
        "tags": [
          "RefTypeRetour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeRetourDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeRetourDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefTypeRetourDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RefTypeRetour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RefTypeRetourDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RefTypeRetour"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RetourFournisseur": {
      "get": {
        "tags": [
          "RetourFournisseur"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RetourFournisseurDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RetourFournisseurDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RetourFournisseurDTO"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RetourFournisseur"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/RetourFournisseur/{id}": {
      "get": {
        "tags": [
          "RetourFournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RetourFournisseurDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetourFournisseurDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetourFournisseurDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "RetourFournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RetourFournisseurDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "RetourFournisseur"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/RetourFournisseur/search": {
      "get": {
        "tags": [
          "RetourFournisseur"
        ],
        "parameters": [
          {
            "name": "societe",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "marque",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reference",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "typeRetour",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDetailsDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDetailsDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FournisseurDetailsDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleDTO": {
        "required": [
          "couleur",
          "description",
          "reference"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "prixDeVenteInter": {
            "type": "number",
            "format": "double"
          },
          "prixDeVenteRevendeur": {
            "type": "number",
            "format": "double"
          },
          "prixDeVenteClientFinal": {
            "type": "number",
            "format": "double"
          },
          "marqueId": {
            "type": "integer",
            "format": "int32"
          },
          "marqueLibelle": {
            "type": "string",
            "nullable": true
          },
          "typeId": {
            "type": "integer",
            "format": "int32"
          },
          "typeLibelle": {
            "type": "string",
            "nullable": true
          },
          "couleur": {
            "type": "string",
            "nullable": true
          },
          "fournisseurId": {
            "type": "integer",
            "format": "int32"
          },
          "fournisseurNom": {
            "type": "string",
            "nullable": true
          },
          "reference": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FournisseurDTO": {
        "required": [
          "adresse",
          "deviseCode",
          "deviseLibelle",
          "ice",
          "nomFournisseur",
          "pays",
          "telephone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nomFournisseur": {
            "type": "string",
            "nullable": true
          },
          "adresse": {
            "type": "string",
            "nullable": true
          },
          "telephone": {
            "type": "string",
            "nullable": true
          },
          "deviseId": {
            "type": "integer",
            "format": "int32"
          },
          "ice": {
            "type": "string",
            "nullable": true
          },
          "pays": {
            "type": "string",
            "nullable": true
          },
          "deviseLibelle": {
            "type": "string",
            "nullable": true
          },
          "deviseCode": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "FournisseurDetailsDTO": {
        "type": "object",
        "properties": {
          "societe": {
            "type": "string",
            "nullable": true
          },
          "fournisseur": {
            "type": "string",
            "nullable": true
          },
          "numFacture": {
            "type": "string",
            "nullable": true
          },
          "numBL": {
            "type": "string",
            "nullable": true
          },
          "dateArrivage": {
            "type": "string",
            "format": "date-time"
          },
          "dateRetour": {
            "type": "string",
            "format": "date-time"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "quantiteRetournee": {
            "type": "integer",
            "format": "int32"
          },
          "marque": {
            "type": "string",
            "nullable": true
          },
          "prixDachatDeviseTTC": {
            "type": "number",
            "format": "double"
          },
          "prixDachatMADTTC": {
            "type": "number",
            "format": "double"
          },
          "motifRetour": {
            "type": "string",
            "nullable": true
          },
          "typeRetour": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RefTypeDTO": {
        "required": [
          "code",
          "libelle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "libelle": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RefTypeRetourDTO": {
        "required": [
          "code",
          "libelle"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "libelle": {
            "type": "string",
            "nullable": true
          },
          "code": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RetourFournisseurDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "arrivageDetailsId": {
            "type": "integer",
            "format": "int32"
          },
          "quantiteRetourne": {
            "type": "integer",
            "format": "int32"
          },
          "motifRetour": {
            "type": "string",
            "nullable": true
          },
          "typeRetourId": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "dateRetour": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}